+++ Conectados2/Models/Persona.cs
@@ -10,7 +10,7 @@ namespace Conectados2.Models
             Patrullero = new HashSet<Patrullero>();
         }
 
-        public long IdPersona { get; set; }
+        public int IdPersona { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string NumDoc { get; set; }
