+++ Conectados2/Models/RolUsuario.cs
@@ -7,8 +7,8 @@ namespace Conectados2.Models
     {
         public int IdRolUsuario { get; set; }
         public string Descripcion { get; set; }
-        public long IdRol { get; set; }
-        public long IdUsuario { get; set; }
+        public int IdRol { get; set; }
+        public int IdUsuario { get; set; }
 
         public Rol IdRolNavigation { get; set; }
         public Usuario IdUsuarioNavigation { get; set; }
